+++ tests/test_treespec.py
@@ -704,7 +704,11 @@ def test_treespec_compose_children(
         stack = [(composed_treespec.children(), expected_treespec.children())]
         while stack:
             composed_children, expected_children = stack.pop()
-            for composed_child, expected_child in zip(composed_children, expected_children):
+            for composed_child, expected_child in zip(
+                composed_children,
+                expected_children,
+                strict=True,
+            ):
                 assert composed_child == expected_child
                 stack.append((composed_child.children(), expected_child.children()))
 
@@ -786,7 +790,7 @@ def test_treespec_entries(
                 yield ()
                 return
 
-            for entry, child in zip(entries, children):
+            for entry, child in zip(entries, children, strict=True):
                 for suffix in gen_path(child):
                     yield (entry, *suffix)
 
@@ -815,7 +819,7 @@ def test_treespec_entries(
 
             node_type = spec.type
             node_kind = spec.kind
-            for entry, child in zip(entries, children):
+            for entry, child in zip(entries, children, strict=True):
                 for suffix in gen_typed_path(child):
                     yield ((entry, node_type, node_kind), *suffix)
 
@@ -1078,11 +1082,11 @@ def test_treespec_transform():
     ) == optree.tree_structure([[1, 2, 3], [4]])
     assert optree.treespec_transform(
         treespec,
-        lambda spec: optree.treespec_dict(zip('abcd', spec.children())),
+        lambda spec: optree.treespec_dict(zip('abcd', spec.children(), strict=False)),
     ) == optree.tree_structure({'a': {'a': 0, 'b': 1, 'c': 2}, 'b': {'a': 3}})
     assert optree.treespec_transform(
         treespec,
-        lambda spec: optree.treespec_dict(zip('abcd', spec.children())),
+        lambda spec: optree.treespec_dict(zip('abcd', spec.children(), strict=False)),
         lambda spec: optree.tree_structure([0, None, 1]),
     ) == optree.tree_structure(
         {'a': {'a': [0, None, 1], 'b': [2, None, 3], 'c': [4, None, 5]}, 'b': {'a': [6, None, 7]}},
@@ -1095,7 +1099,7 @@ def test_treespec_transform():
         optree.treespec_transform(
             treespec,
             lambda spec: optree.tree_structure(
-                MyAnotherDict(zip(spec.entries(), spec.children())),
+                MyAnotherDict(zip(spec.entries(), spec.children(), strict=True)),
                 namespace='namespace',
             ),
         )
@@ -1135,7 +1139,10 @@ def test_treespec_transform():
 
     def fn(spec):
         with optree.dict_insertion_ordered(True, namespace='undefined'):
-            return optree.treespec_dict(zip('abcd', spec.children()), namespace='undefined')
+            return optree.treespec_dict(
+                zip('abcd', spec.children(), strict=False),
+                namespace='undefined',
+            )
 
     with pytest.raises(ValueError, match=r'Expected treespec\(s\) with namespace .*, got .*\.'):
         optree.treespec_transform(namespaced_treespec, fn)
@@ -1556,7 +1563,7 @@ def test_treespec_constructor(  # noqa: C901
                         if dict_should_be_sorted or dict_session_namespace not in {'', namespace}:
                             assert (
                                 optree.treespec_dict(
-                                    zip(sorted(node), children_treespecs),
+                                    zip(sorted(node), children_treespecs, strict=True),
                                     none_is_leaf=none_is_leaf,
                                     namespace=passed_namespace,
                                 )
@@ -1564,7 +1571,7 @@ def test_treespec_constructor(  # noqa: C901
                             )
                             assert (
                                 optree.treespec_from_collection(
-                                    dict(zip(sorted(node), children_treespecs)),
+                                    dict(zip(sorted(node), children_treespecs, strict=True)),
                                     none_is_leaf=none_is_leaf,
                                     namespace=passed_namespace,
                                 )
@@ -1582,7 +1589,7 @@ def test_treespec_constructor(  # noqa: C901
                             with context:
                                 assert (
                                     optree.treespec_dict(
-                                        zip(node, children_treespecs),
+                                        zip(node, children_treespecs, strict=True),
                                         none_is_leaf=none_is_leaf,
                                         namespace=passed_namespace,
                                     )
@@ -1590,7 +1597,7 @@ def test_treespec_constructor(  # noqa: C901
                                 )
                                 assert (
                                     optree.treespec_from_collection(
-                                        dict(zip(node, children_treespecs)),
+                                        dict(zip(node, children_treespecs, strict=True)),
                                         none_is_leaf=none_is_leaf,
                                         namespace=passed_namespace,
                                     )
@@ -1599,7 +1606,7 @@ def test_treespec_constructor(  # noqa: C901
                     elif node_type is OrderedDict:
                         assert (
                             optree.treespec_ordereddict(
-                                zip(node, children_treespecs),
+                                zip(node, children_treespecs, strict=True),
                                 none_is_leaf=none_is_leaf,
                                 namespace=passed_namespace,
                             )
@@ -1607,7 +1614,7 @@ def test_treespec_constructor(  # noqa: C901
                         )
                         assert (
                             optree.treespec_from_collection(
-                                OrderedDict(zip(node, children_treespecs)),
+                                OrderedDict(zip(node, children_treespecs, strict=True)),
                                 none_is_leaf=none_is_leaf,
                                 namespace=passed_namespace,
                             )
@@ -1618,7 +1625,7 @@ def test_treespec_constructor(  # noqa: C901
                             assert (
                                 optree.treespec_defaultdict(
                                     node.default_factory,
-                                    zip(sorted(node), children_treespecs),
+                                    zip(sorted(node), children_treespecs, strict=True),
                                     none_is_leaf=none_is_leaf,
                                     namespace=passed_namespace,
                                 )
@@ -1628,7 +1635,7 @@ def test_treespec_constructor(  # noqa: C901
                                 optree.treespec_from_collection(
                                     defaultdict(
                                         node.default_factory,
-                                        zip(sorted(node), children_treespecs),
+                                        zip(sorted(node), children_treespecs, strict=True),
                                     ),
                                     none_is_leaf=none_is_leaf,
                                     namespace=passed_namespace,
@@ -1648,7 +1655,7 @@ def test_treespec_constructor(  # noqa: C901
                                 assert (
                                     optree.treespec_defaultdict(
                                         node.default_factory,
-                                        zip(node, children_treespecs),
+                                        zip(node, children_treespecs, strict=True),
                                         none_is_leaf=none_is_leaf,
                                         namespace=passed_namespace,
                                     )
@@ -1658,7 +1665,7 @@ def test_treespec_constructor(  # noqa: C901
                                     optree.treespec_from_collection(
                                         defaultdict(
                                             node.default_factory,
-                                            zip(node, children_treespecs),
+                                            zip(node, children_treespecs, strict=True),
                                         ),
                                         none_is_leaf=none_is_leaf,
                                         namespace=passed_namespace,